Delete add_prefixes function
Function purpose was to prefix extra arguments with - or --
For example, if `--args 'quiet-build=n A'` was supplied to
show compile logs and alerts via terminal bell, the add_prefixes
will modify it as `--quiet-build -A` in the runtime.

There is no need for it, as users can just supply already prefixed
arguments, which is more intuitive way.